Access Token Consistency (#106)
* Access Token Consistency

The specification was inconsistent regarding the requirement to present an access token with a request -- the language was inconsistent, and in one case (Delete Request) missing.  For consistency and to eliminate any ambiguity, used the same wording for every operation:  "The request MUST include a valid access token in the Authorization header (i.e., Authorization: Bearer {{token}}). See the Security Considerations section above."

Also, the wording of the description of the "requiresAccessToken" field seemed to minimize the need to include an access token for what is arguably the most security-critical operation -- downloading PHI.  So I added wording that clarifies that if the download is RESTful (e.g., SMART), an access token is required, while acknowledging that some download options (e.g., signed URLs to Amazon S3 buckets) use alternative means of assuring that accesses are authorized.

* Update export.md

Removed two references to the Authorization header.

* Update export.md

Removed references to "access token" except those relating to the requireAccessToken field.