accessTokenRequired definition
Revised wording to clarify when this field MUST hold a value of "true" and when it MAY be "false.  Note slight change in the first sentence from that proposed in the Report, in order to clarify the context of "require OAuth 2.0 bearer tokens."   Also added "bearer" in the first sentence of the definition.